Move method create_symlinks_checkpoint_files
from SimulNek to SimulKTH, because checkpoints are only valid there
